music_theory/requirements.txt
reportlab
music21
Pillow